requirements.txt
streamlit==1.24.0
langchain==0.1.17
langchain-community==0.0.35
langchain-core==0.2.5
langchain-groq==0.0.2
mysql-connector-python==8.0.30
python-dotenv==1.0.0
urllib3==2.0.3
